Remove unused spotify and twitter image domains
This config entry was needed for any next/image components.

We used to display the spotify album art in the NowPlaying
component, but that was removed in
a9534a493c8fb6a8aedb9bddf5f13b43c8d928e3.

We used to display twitter profile pic on /tweets, but I removed that
page in `bba3fa1: Remove twitter`