Overhaul data fetching of posts
The contents of mdx.js were confusing to me. Now I think it's much more
clear what the schema for post data is, and how to get to it.

Another big benefit of this is that now it's easier to include the
banner pic for a post in the ProjectPost. I didn't do that yet though
since the styling sucks as it is.

This allowed me to do some DRY in index.js by being able to load
post data just from a slug, and from
there wew can access whatever metadata is needed

I renamed `frontMatter` to `meta` in a lot of cases because for example
readingTime is a generated meta property,
it's not actually included as the frontMatter in each file.

I removed the `...` operator when passing props to projects/[slug].js,
since I like it better when it's explicit exactly which properties
are needed.

I removed a lot of the `type` arguments from mdx.js functions, since it
looks to me like we only are using it in a single way.
If we need to support other directories later, then I can
add that back in.